backend/pyproject.toml
[tool.poetry]
name = "form-generator"
version = "0.1.0"
description = ""
authors = ["Sergio <sergioricjr7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
pytest = "^8.3.3"
dependency-injector = "^4.42.0"
uvicorn = "^0.31.1"
google-auth = "^2.35.0"
requests = "^2.32.3"
odmantic = "^1.0.2"
aiologger = "^0.7.0"
pymongo = "4.9"
motor = "^3.6.0"
gunicorn = "^23.0.0"
aiokafka = "^0.12.0"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
